+++ components/catalogs/back/src/Infrastructure/Command/CatalogFixtureCommand.php
@@ -197,7 +197,7 @@ class CatalogFixtureCommand extends Command
         return <<<'JSON_WRAP'
         {
           "$id": "https://example.com/product",
-          "$schema": "https://api.akeneo.com/mapping/product/0.0.9/schema",
+          "$schema": "https://api.akeneo.com/mapping/product/0.0.10/schema",
           "$comment": "My first schema !",
           "title": "Product Mapping",
           "description": "JSON Schema describing the structure of products expected by our application",
@@ -250,7 +250,8 @@ class CatalogFixtureCommand extends Command
               "type": "string",
               "enum": ["S", "M", "L"]
             }
-          }
+          },
+          "required": ["name", "size"]
         }
         JSON_WRAP;
     }