+++ components/catalogs/back/tests/Acceptance/ApiContext.php
@@ -751,7 +751,7 @@ class ApiContext implements Context
                 <<<'JSON_WRAP'
                 {
                   "$id": "https://example.com/product",
-                  "$schema": "https://api.akeneo.com/mapping/product/0.0.6/schema",
+                  "$schema": "https://api.akeneo.com/mapping/product/0.0.10/schema",
                   "$comment": "My first schema !",
                   "title": "Product Mapping",
                   "description": "JSON Schema describing the structure of products expected by our application",
@@ -798,7 +798,8 @@ class ApiContext implements Context
                     "type": {
                       "type": "string"
                     }
-                  }
+                  },
+                  "required": ["title"]
                 }
                 JSON_WRAP,
                 false,
@@ -946,6 +947,24 @@ class ApiContext implements Context
                 'sku' => 't-shirt red',
                 'name' => 'T-shirt red',
                 'description' => 'Description red',
+                'size' => 'xl',
+                'drawings_customization_count' => '4',
+                'artists_customization_count' => '2',
+                'price' => ['USD' => 78.3],
+                'released_at' => new \DateTimeImmutable('2042-01-01T00:00:00+00:00'),
+                'is_released' => false,
+                'picture' => $this->files['ziggyImage'],
+                'enabled' => true,
+                'sale_countries' => [
+                    'france',
+                    'brazil',
+                ],
+            ],
+            [
+                'uuid' => '7343e656-a114-4956-bb5e-2f5f1317b6d2',
+                'sku' => 't-shirt yellow',
+                'name' => '',
+                'description' => 'Description yellow',
                 'size' => 'xl',
                 'drawings_customization_count' => '4',
                 'artists_customization_count' => '2',